+++ examples/data-sources/netapp-ontap_volume_efficiency_policy/data-source.tf
@@ -1,0 +1,8 @@
+data "netapp-ontap_volume_efficiency_policy" "volume_efficiency_policy" {
+  # required to know which system to interface with
+  cx_profile_name = "cluster1"
+  name = "test"
+  svm = {
+    name = "terraform"
+  }
+}